Moves PointState out to its own file.
